% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{dutchmasters}
\alias{dutchmasters}
\title{Dutch Masters Colour Palletes}
\format{An object of class \code{list} of length 6.}
\usage{
dutchmasters
}
\description{
Colour palletes derived from famous paintings from the Dutch masters
Johannes Vermeer and Rembrandt van Rijn. Inspired by the \link{ochRe}
(https://github.com/ropenscilabs/ochRe) package.
}
\details{
Available palletes:

Vermeer - The Milkmaid

Vermeer - Girl with a Pearl Earring

Vermeer - View of Delft

Vermeer - The Little Street

Van Rijn - The Anatomy Lesson of Dr. Nicolaes Tulp

Van Rijn - The "Staalmeesters"
}
\examples{
# devtools::install_github("EdwinTh/dutchmasters")
# devtools::install_github("ropenscilabs/ochRe")

gg
}
\keyword{datasets}
