% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{dutchmasters_pal}
\alias{dutchmasters_pal}
\title{dutchmasters palette with ramped colours}
\usage{
dutchmasters_pal(palette = "milkmaid", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'dutchmasters_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
dutchmasters palette with ramped colours
}
\examples{
library(scales)
show_col(dutchmasters_pal("milkmaid")(10))

filled.contour(volcano,color.palette = dutchmasters_pal("pearl_earring"), asp=1)

}
