% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_dutchmasters}
\alias{scale_fill_dutchmasters}
\title{Setup fill palette for ggplot2}
\usage{
scale_fill_dutchmasters(..., palette = "milkmaid", discrete = TRUE,
  alpha = 1, reverse = TRUE)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'dutchmasters_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
Setup fill palette for ggplot2
}
